Changed size_t to int for im2col and col2im kernels and C++ code

+++ src/ocl/convolutionocl.cpp
@@ -364,7 +364,7 @@ void ConvolutionDescriptor::FindConvFwdAlgorithm(Handle& handle,
                 MIOPEN_LOG_FUNCTION("transpose, non 1x1");
 
                 float time_col2im = 0;
-                size_t out_offset = 0;
+                int out_offset = 0;
 
                 gg.FindSolution(.003, handle, w, x, workSpace, false);
                 gg.RunGemm(handle, w, x, workSpace, 0, 0, 0);
@@ -477,7 +477,7 @@ void ConvolutionDescriptor::FindConvFwdAlgorithm(Handle& handle,
                 GemmDescriptor gemm_desc = CreateGemmDescriptorConvFwd(wDesc, xDesc, yDesc);
 
                 float time_im2col = 0;
-                size_t in_offset  = 0;
+                int in_offset  = 0;
                 time_im2col       = Im2ColGPU(handle,
                                         xDesc.GetElementSize(),
                                         x,
@@ -942,7 +942,7 @@ void ConvolutionDescriptor::ConvolutionForward(Handle& handle,
                 for(int i = 0; i < in_n; i++)
                 {
                     int out_offset   = i * wei_n * out_h * out_w;
-                    size_t in_offset = i * in_c * in_h * in_w;
+                    int in_offset = i * in_c * in_h * in_w;
                     Im2ColGPU(handle,
                               xDesc.GetElementSize(),
                               x,
@@ -1045,7 +1045,7 @@ void ConvolutionDescriptor::ConvolutionForward(Handle& handle,
             {
                 MIOPEN_LOG_FUNCTION("transppose, non 1x1");
 
-                size_t in_offset = i * in_c * in_h * in_w;
+                int in_offset = i * in_c * in_h * in_w;
 
                 gg.RunGemm(handle, w, x, workSpace, 0, in_offset, 0);
 
@@ -1182,7 +1182,7 @@ void ConvolutionDescriptor::FindConvBwdDataAlgorithm(Handle& handle,
                 MIOPEN_LOG_FUNCTION("transppose, non 1x1");
 
                 float time_im2col = 0;
-                size_t out_offset = 0;
+                int out_offset = 0;
                 time_im2col       = Im2ColGPU(handle,
                                         dyDesc.GetElementSize(),
                                         dy,
@@ -1445,7 +1445,7 @@ void ConvolutionDescriptor::FindConvBwdDataAlgorithm(Handle& handle,
                 GemmDescriptor gemm_desc = CreateGemmDescriptorConvBwdData(wDesc, dyDesc, dxDesc);
 
                 float time_col2im = 0;
-                size_t in_offset  = 0;
+                int in_offset  = 0;
 
                 // dx = transpose(w) * dy
                 CallGemm(handle, gemm_desc, w, 0, dy, 0, workSpace, 0);
@@ -1787,7 +1787,7 @@ void ConvolutionDescriptor::ConvolutionBackwardData(Handle& handle,
                 for(int i = 0; i < in_n; i++)
                 {
                     int out_offset   = i * wei_n * out_h * out_w;
-                    size_t in_offset = i * in_c * in_h * in_w;
+                    int in_offset = i * in_c * in_h * in_w;
 
                     // dx = transpose(w) * dy
                     CallGemm(handle, gemm_desc, w, 0, dy, out_offset, workSpace, 0);
@@ -1889,7 +1889,7 @@ void ConvolutionDescriptor::ConvolutionBackwardData(Handle& handle,
             {
                 MIOPEN_LOG_FUNCTION("transpose, non 1x1");
 
-                size_t out_offset = i * wei_n * out_h * out_w;
+                int out_offset = i * wei_n * out_h * out_w;
                 Im2ColGPU(handle,
                           dyDesc.GetElementSize(),
                           dy,
@@ -2093,7 +2093,7 @@ void ConvolutionDescriptor::FindConvBwdWeightsAlgorithm(Handle& handle,
                 MIOPEN_LOG_FUNCTION("transpose, non 1x1");
 
                 float time_im2col = 0;
-                size_t out_offset = 0;
+                int out_offset = 0;
                 time_im2col       = Im2ColGPU(handle,
                                         dyDesc.GetElementSize(),
                                         dy,
@@ -2143,7 +2143,7 @@ void ConvolutionDescriptor::FindConvBwdWeightsAlgorithm(Handle& handle,
                 GemmDescriptor gemm_desc = CreateGemmDescriptorConvBwdWeight(dyDesc, xDesc, dwDesc);
 
                 float time_im2col = 0;
-                size_t in_offset  = 0;
+                int in_offset  = 0;
                 time_im2col       = Im2ColGPU(handle,
                                         xDesc.GetElementSize(),
                                         x,
@@ -2358,7 +2358,7 @@ void ConvolutionDescriptor::ConvolutionBackwardWeights(Handle& handle,
                 for(int i = 0; i < in_n; i++)
                 {
                     int out_offset   = i * wei_n * out_h * out_w;
-                    size_t in_offset = i * in_c * in_h * in_w;
+                    int in_offset = i * in_c * in_h * in_w;
                     Im2ColGPU(handle,
                               xDesc.GetElementSize(),
                               x,
@@ -2543,7 +2543,7 @@ void ConvolutionDescriptor::ConvolutionBackwardWeights(Handle& handle,
             {
                 MIOPEN_LOG_FUNCTION("transpose, non 1x1");
 
-                size_t out_offset = i * wei_n * out_h * out_w;
+                int out_offset = i * wei_n * out_h * out_w;
                 Im2ColGPU(handle,
                           dyDesc.GetElementSize(),
                           dy,
